Improve the handling of multiple downlink interfaces.
This patch has two effects. First, the quickleave feature/behaviour is disabled for all groups that are used on more than one interface. The idea of quickleave is to leave a group fast and later figure out whether there is still somebody interested in that group. For groups used on more than one interface, it is already known that there is still somebody interested in that group.

Second, when a leave is received for a group that is used on more than one interface, igmpproxy sends queries on all interface to discover remeining listeners for that group. Previously these queries were only send on the interface the leave was received on, so that listeners on the other interfaces were not discovered and the group might be left on the upstream router incorrectly.

This patch can be improved by sending the queries only on the interface the leave was received on and adapting the algorithm in internAgeRoute(...) in rttable.c in a way that only one interface is actually processed and all other interfaces of the route are silently assumed to be still active.

This patch has also been submitted to the upstream maintainer: https://sourceforge.net/p/igmpproxy/patches/9/

It is also available to pull for the upstream maintainer on github: https://github.com/eriktews/igmpproxy/commit/f0b483a2a02574c1e0ffde68b227c8c8bd689a7e